# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing commands, tokens, and access logs",
        "title": "httprun API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api",
    "paths": {
...
        "/run": {
            "post": {
                "description": "Run a command if the user has permission",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "commands"
                ],
                "summary": "Run a command",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT token",
                        "name": "x-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Command object to run",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.RunCommandRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Command executed successfully",
                        "schema": {
                            "$ref": "#/definitions/types.CommandResponse"
                        }
                    }
                }
            }
        },
        "/run/commands": {
            "get": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "description": "Lists commands that the user has permission to run",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "commands"
                ],
                "summary": "List commands",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Command"
                            }
                        }
                    },
                    "403": {
                        "description": "Permission denied",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "command.Env": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "command.Param": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "value": {}
            }
        },
...
        "models.Command": {
            "type": "object",
            "properties": {
                "command": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.CommandStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.CommandStatus": {
            "type": "integer",
            "enum": [
                0,
                1
            ],
            "x-enum-varnames": [
                "CommandStatusActive",
                "CommandStatusInactive"
            ]
        },
...
        "types.CommandResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "stderr": {
                    "type": "string"
                },
                "stdout": {
                    "type": "string"
                }
            }
        },
...
        "types.RunCommandRequest": {
            "type": "object",
            "properties": {
                "env": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/command.Env"
                    }
                },
                "name": {
                    "type": "string"
                },
                "params": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/command.Param"
                    }
                }
            }
        },
...
    }
}